Created dedicated unit test for Multi-Curve Calibration.
